+++ src/elements/actions/Translate.php
@@ -29,7 +29,7 @@ class Translate extends ElementAction
         type: $type,
         bulk: true,
         // Return whether the action should be available depending on which elements are selected
-        validateSelection: (selectedItems) {
+        validateSelection: (selectedItems) => {
           return true;
         },
     });