+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Release Notes for Multi Translator
 
+## 1.12.0 - 2024-08-13
+
+### Added
+
+- added support for verbb\hyper\fields\HyperField
+- more openai models in settings list
+
+### Fixed
+
+- fix js syntax when registering action trigger
+- only translate title when translatable
+- use temperature setting in openai body
+
 ## 1.11.0 - 2024-07-09
 
 ### Added